% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{peptide_identification}
\alias{peptide_identification}
\title{Title}
\usage{
peptide_identification(x, peptideSequence,
  pimIdx = parentIonMass(peptideSequence), peptideMassTolerancePPM = 5,
  framentIonMassToleranceDa = 0.01, FUN = .byIon)
}
\arguments{
\item{x}{a mass spectrometric measurement}

\item{peptideSequence}{digested, e.g. tryptic, peptide sequnences, sorted according to the parent ion mass}

\item{pimIdx}{parent ion mass index}

\item{peptideMassTolerancePPM}{}

\item{framentIonMassToleranceDa}{}

\item{FUN}{}
}
\value{
a data.frame
}
\description{
Title
}

